+++ CHANGES.md
@@ -19,3 +19,5 @@
 ## Learning-JavaScript-Course-Task
  - Applied the YAGNI principle
    - Removed the improvements that wasn't in the task
+ - Applied the SOLID principle
+   - Refactored code in [Learning-JavaScript-Course-Task.optimized/task2/script.js](Learning-JavaScript-Course-Task.optimized/task2/script.js) according to the Open/Closed principle